pyproject.toml
[project]
name = "amplifier-module-image-generation"
version = "0.1.0"
description = "Multi-provider AI image generation module for Amplifier"
authors = [
    {name = "Brian Krabach", email = "briankrabach@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
]

[project.entry-points."amplifier.tool"]
image-generation = "amplifier_module_image_generation:ImageGenerationTool"

[project.optional-dependencies]
tool = [
    "amplifier-core>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pillow>=12.1.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_module_image_generation"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
include = ["src", "tests"]
exclude = [".venv", "__pycache__", "*.egg-info"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
